Fix segmentation fault
Device holds a shared smart pointer to Brg named `brg` and a shared smart
pointer to STLinkInterface named `stlink`. However, the `brg` member will
end up with a raw pointer to the `stlink` as well. Since the order in which
C++ calls class member destructors is in the reverse order in which they
appear in the class definition, `stlink`'s destructor will be called first,
and then `brg`'s descructor will be called. This leaves `brg` with a dangling
pointer that can lead to a segmentation fault. The solution is to define
`stlink` before `brg` in the Device class definition.